Rewrite create_game presenter and events
