+++ love_letter/models/event.py
@@ -1,2 +1,10 @@
+from dataclasses import dataclass
+
+
 class DomainEvent:
     pass
+
+
+@dataclass
+class GameCreatedEvent(DomainEvent):
+    game_id: str
